Create traits for the config, identity, and event APIs
