+++ libraries/joomla/application/base.php
@@ -10,10 +10,10 @@
 defined('JPATH_PLATFORM') or die;
 
 use Joomla\Application\AbstractApplication;
+use Joomla\Cms\Application\EventAware;
+use Joomla\Cms\Application\IdentityAware;
 use Joomla\Event\DispatcherAwareInterface;
 use Joomla\Event\DispatcherAwareTrait;
-use Joomla\Event\DispatcherInterface;
-use Joomla\Event\Event;
 use Joomla\Registry\Registry;
 
 /**
@@ -25,15 +25,7 @@ use Joomla\Registry\Registry;
  */
 abstract class JApplicationBase extends AbstractApplication implements DispatcherAwareInterface
 {
-	use DispatcherAwareTrait;
-
-	/**
-	 * The application identity object.
-	 *
-	 * @var    JUser
-	 * @since  12.1
-	 */
-	protected $identity;
+	use DispatcherAwareTrait, EventAware, IdentityAware;
 
 	/**
 	 * Class constructor.
@@ -56,108 +48,6 @@ abstract class JApplicationBase extends AbstractApplication implements Dispatche
 	}
 
 	/**
-	 * Get the application identity.
-	 *
-	 * @return  mixed  A JUser object or null.
-	 *
-	 * @since   12.1
-	 */
-	public function getIdentity()
-	{
-		return $this->identity;
-	}
-
-	/**
-	 * Registers a handler to a particular event group.
-	 *
-	 * @param   string    $event    The event name.
-	 * @param   callable  $handler  The handler, a function or an instance of an event object.
-	 *
-	 * @return  JApplicationBase  The application to allow chaining.
-	 *
-	 * @since   12.1
-	 */
-	public function registerEvent($event, $handler)
-	{
-		try
-		{
-			$this->getDispatcher()->addListener($event, $handler);
-		}
-		catch (UnexpectedValueException $e)
-  		{
-			// No dispatcher is registered, don't throw an error (mimics old behavior)
-  		}
-
-		return $this;
-	}
-
-	/**
-	 * Calls all handlers associated with an event group.
-	 *
-	 * This is a legacy method, implementing old-style (Joomla! 3.x) plugin calls. It's best to go directly through the
-	 * Dispatcher and handle the returned EventInterface object instead of going through this method. This method is
-	 * deprecated and will be removed in Joomla! 5.x.
-	 *
-	 * This method will only return the 'result' argument of the event
-	 *
-	 * @param   string        $eventName  The event name.
-	 * @param   array|Event   $args       An array of arguments or an Event object (optional).
-	 *
-	 * @return  array   An array of results from each function call, or null if no dispatcher is defined.
-	 *
-	 * @since       12.1
-	 * @throws      InvalidArgumentException
-	 * @deprecated  5.0
-	 */
-	public function triggerEvent($eventName, $args = array())
-	{
-		$dispatcher = $this->getDispatcher();
-
-		if ($this->dispatcher instanceof DispatcherInterface)
-		{
-			if ($args instanceof Event)
-			{
-				$event = $args;
-			}
-			elseif (is_array($args))
-			{
-				$event = new Event($eventName, $args);
-			}
-			else
-			{
-				throw new InvalidArgumentException('The arguments must either be an event or an array');
-			}
-
-			$result = $dispatcher->dispatch($eventName, $event);
-
-			// TODO - There are still test cases where the result isn't defined, temporarily leave the isset check in place
-			return !isset($result['result']) || is_null($result['result']) ? [] : $result['result'];
-		}
-
-		return;
-	}
-
-	/**
-	 * Allows the application to load a custom or default identity.
-	 *
-	 * The logic and options for creating this object are adequately generic for default cases
-	 * but for many applications it will make sense to override this method and create an identity,
-	 * if required, based on more specific needs.
-	 *
-	 * @param   JUser  $identity  An optional identity object. If omitted, the factory user is created.
-	 *
-	 * @return  JApplicationBase This method is chainable.
-	 *
-	 * @since   12.1
-	 */
-	public function loadIdentity(JUser $identity = null)
-	{
-		$this->identity = ($identity === null) ? JFactory::getUser() : $identity;
-
-		return $this;
-	}
-
-	/**
 	 * Method to run the application routines.  Most likely you will want to instantiate a controller
 	 * and execute it, or perform some sort of task directly.
 	 *
@@ -170,5 +60,4 @@ abstract class JApplicationBase extends AbstractApplication implements Dispatche
 	{
 		return;
 	}
-
 }
